AssetManagementSystem/src/main/java/com/sprint1/assetmanagementsystem/service/ShipmentService.java
package com.sprint1.assetmanagementsystem.service;


import java.util.List;

import com.sprint1.assetmanagementsystem.exception.LoginException;
import com.sprint1.assetmanagementsystem.exception.NoAccessException;
import com.sprint1.assetmanagementsystem.exception.NotFoundException;
import com.sprint1.assetmanagementsystem.model.Shipment;

public interface ShipmentService {
	public Shipment createShipment(Shipment shipment)throws NotFoundException;
	public List<Shipment> viewShipment()throws NotFoundException;
	public Shipment getShipmentById(int shipmentId)throws NotFoundException;
	public String deleteShipment(int shipmentId) throws NotFoundException;
}
